backend/requirements.txt
fastapi==0.135.1
uvicorn[standard]==0.41.0
python-dotenv==1.2.2
sqlalchemy==2.0.48
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.5.0
python-multipart==0.0.22
pydantic[email]==2.12.5
google-genai>=1.0.0,<2.0.0
psycopg[binary]>=3.2.0,<4.0.0
PyPDF2==3.0.1
APScheduler==3.11.2
bcrypt==4.0.1
